+++ src/main/kotlin/io/openfuture/chain/exception/SeedValidationException.kt
@@ -1,3 +1,0 @@
-package io.openfuture.chain.exception
-
-class SeedValidationException(message: String) : Exception(message)