+++ src/test/kotlin/io/openfuture/chain/controller/CryptoControllerTests.kt
@@ -3,11 +3,11 @@ package io.openfuture.chain.controller
 import io.openfuture.chain.config.ControllerTests
 import io.openfuture.chain.config.any
 import io.openfuture.chain.crypto.domain.ExtendedKey
-import io.openfuture.chain.domain.crypto.DerivationKeyRequest
-import io.openfuture.chain.domain.crypto.MasterKeyRequest
-import io.openfuture.chain.domain.crypto.ValidateAddressRequest
-import io.openfuture.chain.domain.crypto.key.AddressKeyDto
+import io.openfuture.chain.domain.crypto.AccountDto
+import io.openfuture.chain.domain.crypto.RootAccountDto
+import io.openfuture.chain.domain.crypto.key.DerivationKeyRequest
 import io.openfuture.chain.domain.crypto.key.KeyDto
+import io.openfuture.chain.domain.crypto.key.RestoreRequest
 import io.openfuture.chain.service.CryptoService
 import org.junit.Test
 import org.mockito.BDDMockito.given
@@ -23,32 +23,26 @@ class CryptoControllerTests : ControllerTests() {
 
 
     @Test
-    fun doGenerateMasterReturnMasterKeyWhenSeedPhraseSent() {
-        val seedPhrase = "1 2 3 4 5 6 7 8 9 10 11 12"
-        val seed = ByteArray(32)
-        val masterKey = ExtendedKey.root(seed)
-        val masterKeyRequest = MasterKeyRequest(seedPhrase)
-        val expectedAddress = AddressKeyDto("1", "2")
+    fun doRestoreShouldReturnRootAccountInfoWhenSeedPhraseSent() {
+        val accountDto = createAccountDto()
+        val expectedAccount = RootAccountDto("1 2 3 4 5 6 7 8 9 10 11 12", accountDto.keys, accountDto)
 
-        given(cryptoService.serializePublicKey(any(ExtendedKey::class.java))).willReturn("1")
-        given(cryptoService.serializePrivateKey(any(ExtendedKey::class.java))).willReturn("2")
-        given(cryptoService.getMasterKey(seedPhrase)).willReturn(masterKey)
+        given(cryptoService.getRootAccount(expectedAccount.seedPhrase)).willReturn(expectedAccount)
 
-        webClient.post().uri("${PathConstant.RPC}/crypto/doGenerateMaster")
-            .body(Mono.just(masterKeyRequest), MasterKeyRequest::class.java)
+        webClient.post().uri("${PathConstant.RPC}/crypto/doRestore")
+            .body(Mono.just(RestoreRequest(expectedAccount.seedPhrase)), RestoreRequest::class.java)
             .exchange()
             .expectStatus().isOk
-            .expectBody(AddressKeyDto::class.java).isEqualTo<Nothing>(expectedAddress)
+            .expectBody(RootAccountDto::class.java).isEqualTo<Nothing>(expectedAccount)
     }
 
     @Test
     fun doDeriveReturnDerivationKeyWhenSeedPhraseDerivationPathAndSent() {
         val seedPhrase = "1 2 3 4 5 6 7 8 9 10 11 12"
         val derivationPath = "m/0"
-        val seed = ByteArray(32)
-        val masterKey = ExtendedKey.root(seed)
+        val masterKey = ExtendedKey.root(ByteArray(32))
         val derivationKeyRequest = DerivationKeyRequest(seedPhrase, derivationPath)
-        val expectedKey = KeyDto("1", "2")
+        val expectedAccount = createAccountDto()
 
         given(cryptoService.serializePublicKey(any(ExtendedKey::class.java))).willReturn("1")
         given(cryptoService.serializePrivateKey(any(ExtendedKey::class.java))).willReturn("2")
@@ -58,30 +52,10 @@ class CryptoControllerTests : ControllerTests() {
             .body(Mono.just(derivationKeyRequest), DerivationKeyRequest::class.java)
             .exchange()
             .expectStatus().isOk
-            .expectBody(KeyDto::class.java).isEqualTo<Nothing>(expectedKey)
+            .expectBody(AccountDto::class.java).isEqualTo<Nothing>(expectedAccount)
     }
 
-    @Test
-    fun validateAddressShouldReturnAddressAndStatusOk() {
-        val address = "0x5aF3B0FFB89C09D7A38Fd01E42E0A5e32011e36e"
-        val request = ValidateAddressRequest(address)
-        
-        webClient.post().uri("${PathConstant.RPC}/crypto/validateAddress")
-            .body(Mono.just(request), ValidateAddressRequest::class.java)
-            .exchange()
-            .expectStatus().isOk
-            .expectBody(ValidateAddressRequest::class.java).isEqualTo<Nothing>(request)
-    }
-
-    @Test
-    fun validateAddressShouldReturnStatusBadRequest() {
-        val address = "0x5aF3B0FFB89C09D7A38Fd01E42E0A5e32011e36eaaaa"
-        val request = ValidateAddressRequest(address)
-
-        webClient.post().uri("${PathConstant.RPC}/crypto/validateAddress")
-            .body(Mono.just(request), ValidateAddressRequest::class.java)
-            .exchange()
-            .expectStatus().isBadRequest
-    }
+    private fun createAccountDto(): AccountDto =
+        AccountDto(KeyDto("1", "2"), "0x83a1e77Bd25daADd7A889BC36AC207A7D39CFD02")
 
 }